requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.2.5
numpy==1.18.1
pandas==1.0.3
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.1
xlrd==1.2.0
